requirements.txt
annotated-doc==0.0.4
annotated-types==0.7.0
anthropic==0.79.0
anyio==4.12.1
bcrypt==5.0.0
blinker==1.9.0
certifi==2026.1.4
charset-normalizer==3.4.4
click==8.3.1
distro==1.9.0
docstring_parser==0.17.0
fastapi==0.135.2
Flask==3.1.3
gunicorn==25.2.0
h11==0.16.0
httpcore==1.0.9
httptools==0.7.1
httpx==0.28.1
idna==3.11
itsdangerous==2.2.0
Jinja2==3.1.6
jiter==0.13.0
markdown-it-py==4.0.0
MarkupSafe==3.0.3
mdurl==0.1.2
mysql-connector-python==9.6.0
numpy==2.4.2
openai==2.21.0
packaging==26.0
pillow==12.1.1
pydantic==2.12.5
pydantic_core==2.41.5
Pygments==2.19.2
python-dotenv==1.2.1
python-telegram-bot==22.6
PyYAML==6.0.3
reportlab==4.4.10
requests==2.32.5
rich==14.3.2
shellingham==1.5.4
sniffio==1.3.1
starlette==1.0.0
tqdm==4.67.3
typer==0.23.1
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.6.3
uvicorn==0.42.0
uvloop==0.22.1
watchfiles==1.1.1
websockets==16.0
Werkzeug==3.1.7
